Fix method invocation
The method takes 2 arguments, but it was only called with one. It
invokes `writeNMD`, so there's no need to invoke it a second time.

An alternative form of this might *not* take the nmd_file as input and
instead leave the Snakefile to write it, but this seems to be a mix of
both.